+++ mobility_features/CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.5] - Private constructor.
+* The Mobility Context constructor is now private
+* A Mobility Context should always be instantiated via the ContextGenerator class.
 
-* TODO: Describe initial release.
+## [0.1.0] - First release.
+* The first official release with working unit tests
+* Includes a minimalistic API which allows the application programmer to generate features with very few lines of code.
+